inline execute.Execute
Now that execute.Execute is only used by SearchClient.Execute, inline it
and delete the execute package.
